data-platform/apps/document-processor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pydantic==2.10.4
pydantic-settings==2.7.0
aiosqlite==0.21.0
chromadb==0.6.3
sentence-transformers==3.3.1
torch==2.5.1
torchvision==0.20.1
Pillow==11.1.0
PyMuPDF==1.25.1
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
httpx==0.28.1
mcp>=1.2.0
openai>=1.0.0
rank-bm25>=0.2.2
pytest==8.3.4
pytest-asyncio==0.25.2
ruff==0.9.1
einops
